Added Derby database support
A new Derby database client has been added to the list of supported databases in the Db.Type enum in ClientRepo.kt. The DerbyClient class has been implemented under the same package and all necessary methods have been overridden as per the requirements of the Derby database. A few updates are also done in build.gradle, and Client.kt to reflect this new addition.